+++ status_map/__init__.py
@@ -3,7 +3,7 @@ from functools import total_ordering
 
 from .exceptions import RepeatedTransition, StatusNotFound, TransitionNotFound
 
-__version__ = "0.2.0"
+__version__ = "0.4.0"
 
 
 @total_ordering
@@ -65,6 +65,8 @@ class StatusMap(Mapping):
         self._transitions = dict(transitions)
         assert self._transitions, "must pass a non-empty dict"
 
+        self.is_cycle = self._is_cycle()
+
         self._statuses = {}
         self._parent = None
 
@@ -97,8 +99,9 @@ class StatusMap(Mapping):
         self._statuses[status.name] = status
 
         for current in status.next:
-            current._add_previous(status)
             current._add_next(self._transitions[current.name])
+            if not self.is_cycle:
+                current._add_previous(status)
 
             if current.name not in self._statuses:
                 self._add_status(current, status)
@@ -121,3 +124,14 @@ class StatusMap(Mapping):
             raise RepeatedTransition(msg)
 
         raise TransitionNotFound(f"transition from {from_status.name} to {to_status.name} not found")
+
+    def _is_cycle(self):
+        visited = set()
+        for status in self._transitions:
+            visited.add(status)
+            for inner_status, values in self._transitions.items():
+                if inner_status in visited:
+                    continue
+                if status in values:
+                    return True
+        return False